refactor(rpc): isolate dispatch path
